+++ src/spb-proto-compiler/parser/char_stream.h
